+++ examples/src/bin/embassy_i2s_read.rs
@@ -12,11 +12,10 @@
 //! You can also inspect the MCLK, BCLK and WS with a logic analyzer.
 
 //% CHIPS: esp32 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: async embassy embassy-executor-thread embassy-time-timg0 embassy-generic-timers
+//% FEATURES: async embassy embassy-time-timg0 embassy-generic-timers
 
 #![no_std]
 #![no_main]
-#![feature(type_alias_impl_trait)]
 
 use embassy_executor::Spawner;
 use esp_backtrace as _;