+++ examples/src/bin/embassy_parl_io_rx.rs
@@ -4,11 +4,10 @@
 //! Uses GPIO 1, 2, 3 and 4 as the data pins.
 
 //% CHIPS: esp32c6 esp32h2
-//% FEATURES: async embassy embassy-executor-thread embassy-time-timg0 embassy-generic-timers
+//% FEATURES: async embassy embassy-time-timg0 embassy-generic-timers
 
 #![no_std]
 #![no_main]
-#![feature(type_alias_impl_trait)]
 
 use embassy_executor::Spawner;
 use embassy_time::{Duration, Timer};