+++ examples/src/bin/embassy_systimer_delay.rs
@@ -3,11 +3,10 @@
 //! This is an example of using the `DelayNs` trait implementation
 
 //% CHIPS: esp32c2 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: embassy embassy-time-timg0 embassy-executor-thread embassy-generic-timers async
+//% FEATURES: embassy embassy-time-timg0 embassy-generic-timers async
 
 #![no_std]
 #![no_main]
-#![feature(type_alias_impl_trait)]
 
 use embassy_executor::Spawner;
 use embedded_hal_async::delay::DelayNs;