+++ hil-test/tests/uart_async.rs
@@ -47,7 +47,7 @@ impl Context {
 }
 
 #[cfg(test)]
-#[embedded_test::tests(executor = esp_hal::embassy::executor::thread::Executor::new())]
+#[embedded_test::tests(executor = esp_hal::embassy::executor::Executor::new())]
 mod tests {
     use defmt::assert_eq;